+++ app/Service/Tag/DevelopmentListService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\Tag;
 
 use App\Models\Resume;
